lab 9

+++ lab-9/nft-staking/programs/nft-staking/src/state/user_account.rs
@@ -1,12 +1,9 @@
-#![allow(unexpected_cfgs)]
-#![allow(deprecated)] // optional, silences internal realloc warning
-
 use anchor_lang::prelude::*;
 
 #[account]
 #[derive(InitSpace)]
 pub struct UserAccount {
     pub points: u32,
-    pub amount_staked: u64,
+    pub amount_staked: u8,
     pub bump: u8,
 }
